Nudge this forwards a bit.
I think I've got the wrong underlying model for menu item positions. It's too rigidly structured with MenuRows and MenuItems. I need to be able to transition items individually so I probably want to do away with MenuRow as a layout concept.

I need to be able to move items vertically individually for dragging into a submenu. I don't want the complexity of having nested objects with different y offsets (and possibly different scales if the selected item doesn't shrink as fast).